// src/action.config.type.ts
//GENERATED FILE FROM report.config.tests.ts - DO NOT EDIT!!!

export type ConfigurationInputs = {
  ShowAdditions: string | number
  ShowDeleted: string | number
  ShowNumberOfChangedFiles: string | number
  ShowNumberOfCommits: string | number
  ShowNumberOfReviews: string | number
  ShowNumberOfComments: string | number
  ShowNumberOfCommentOnlyReviews: string | number
  ShowNumberOfRequestedChangeReviews: string | number
  ShowNumberOfApprovedReviews: string | number
  ShowTotalNumberOfParticipants: string | number
  ShowTimeTotalRuntimeForLastStatusCheckRun: string | number
  ShowTimeSpendOnPrForLastStatusCheckRun: string | number
  ShowPRLeadTime: string | number
  ShowTimeSpendOnBranchBeforePrCreated: string | number
  ShowTimeSpendOnBranchBeforePrMerged: string | number
  ShowTimeToMergeAfterLastReview: string | number
}